Split dashboard browser and style assets
