+++ src/backend/webui-dashboard-browser-injected-helpers.ts
@@ -1,0 +1,146 @@
+import {
+  buildAttentionItems,
+  buildNextIssueSummary,
+  buildOverviewSummary,
+  buildPrimaryActionSummary,
+  describeLoopOffTrackedWorkRestartActionTitle,
+  describeLoopOffTrackedWorkBlocker,
+  describeLoopOffTrackedWorkRestartExpectation,
+  describeLoopOffTrackedWorkRestartGuidance,
+  collectTrackedIssues,
+  collectIssueShortcuts,
+  describeCommandSelectionChange,
+  formatBlockedIssues,
+  formatCandidateDiscovery,
+  formatIssueRef,
+  formatRecentPhaseChanges,
+  formatRecoveryLoopSummary,
+  formatRetryContextSummary,
+  formatRunnableIssues,
+  formatTrackedHistorySummary,
+  formatTrackedIssueSummary,
+  formatTrackedIssues,
+  describeConnectionHealth,
+  describeFreshnessState,
+  describeTimelineCommandResult,
+  describeTimelineEvent,
+  collectTimelineEventIssueNumbers,
+  humanizeTimelineValue,
+  parseSelectedIssueNumber,
+  parseRenderedOperatorAction,
+  readOperatorActionToken,
+  selectRenderedOperatorAction,
+} from "./webui-dashboard-browser-logic";
+import {
+  buildWorkflowSteps,
+  buildRuntimeRecoverySummaryLines,
+  countCandidateIssues,
+  describeLoopRuntime,
+  formatKeyValueBlock,
+  formatRefreshTime,
+  liveBadgeClass,
+  metricClass,
+} from "./webui-dashboard-browser-view-model";
+import {
+  buildIssueExplainSections,
+  formatIssueNumber,
+  formatIssueTimelineEvent,
+  formatIssueTimelineEvents,
+  formatIssueTimelineSummary,
+  formatLatestRecovery,
+  formatReviewWaits,
+} from "./webui-dashboard-browser-issue-details";
+import {
+  createDashboardControlLayer,
+} from "./webui-dashboard-browser-controls";
+import {
+  registerDashboardDomInteractions,
+  wireDashboardEventStream,
+} from "./webui-dashboard-browser-interactions";
+import {
+  formatBrowserToken,
+  normalizeBrowserLocalCiContract,
+  buildBrowserLocalCiStatusLines,
+  buildMutationHeaders,
+  postMutationJsonWithAuth,
+  promptForMutationAuthToken,
+  readMutationResponsePayload,
+  readStoredMutationAuthToken,
+  writeStoredMutationAuthToken,
+} from "./webui-browser-script-helpers";
+import {
+  parseOperatorActionPriority,
+  validOperatorActions,
+} from "../operator-actions";
+
+const injectedBrowserConstants = [
+  `const validOperatorActions = ${JSON.stringify(validOperatorActions)};`,
+].join("\n\n");
+
+const injectedBrowserLogic = [
+  formatBrowserToken,
+  normalizeBrowserLocalCiContract,
+  buildBrowserLocalCiStatusLines,
+  parseOperatorActionPriority,
+  readStoredMutationAuthToken,
+  writeStoredMutationAuthToken,
+  promptForMutationAuthToken,
+  buildMutationHeaders,
+  readMutationResponsePayload,
+  postMutationJsonWithAuth,
+  createDashboardControlLayer,
+  registerDashboardDomInteractions,
+  wireDashboardEventStream,
+  buildOverviewSummary,
+  buildNextIssueSummary,
+  buildPrimaryActionSummary,
+  buildAttentionItems,
+  describeLoopOffTrackedWorkRestartActionTitle,
+  describeLoopOffTrackedWorkRestartExpectation,
+  describeLoopOffTrackedWorkRestartGuidance,
+  describeLoopOffTrackedWorkBlocker,
+  collectTrackedIssues,
+  formatTrackedIssues,
+  formatTrackedHistorySummary,
+  formatTrackedIssueSummary,
+  formatRunnableIssues,
+  formatBlockedIssues,
+  formatCandidateDiscovery,
+  formatIssueRef,
+  formatRetryContextSummary,
+  formatRecoveryLoopSummary,
+  formatRecentPhaseChanges,
+  collectIssueShortcuts,
+  describeCommandSelectionChange,
+  describeConnectionHealth,
+  describeFreshnessState,
+  humanizeTimelineValue,
+  describeTimelineCommandResult,
+  describeTimelineEvent,
+  collectTimelineEventIssueNumbers,
+  parseSelectedIssueNumber,
+  readOperatorActionToken,
+  parseRenderedOperatorAction,
+  selectRenderedOperatorAction,
+  countCandidateIssues,
+  buildWorkflowSteps,
+  buildRuntimeRecoverySummaryLines,
+  metricClass,
+  formatKeyValueBlock,
+  liveBadgeClass,
+  formatRefreshTime,
+  describeLoopRuntime,
+  formatLatestRecovery,
+  formatReviewWaits,
+  formatIssueNumber,
+  formatIssueTimelineEvent,
+  formatIssueTimelineSummary,
+  formatIssueTimelineEvents,
+  buildIssueExplainSections,
+]
+  .map((helper) => helper.toString().replace(/__name\([^;]+;\s*/gu, ""))
+  .join("\n\n");
+
+export function renderDashboardInjectedHelperScript(): string {
+  return [injectedBrowserConstants, injectedBrowserLogic].join("\n\n");
+}
