+++ src/backend/webui-dashboard-browser-state-assets.ts
@@ -1,0 +1,107 @@
+import { WEBUI_MUTATION_AUTH_HEADER, WEBUI_MUTATION_AUTH_STORAGE_KEY } from "./webui-mutation-auth";
+
+export function renderDashboardBrowserStateScript(): string {
+  return `      const state = {
+        selectedIssueNumber: null,
+        loadedIssueNumber: null,
+        status: null,
+        doctor: null,
+        explain: null,
+        issueLint: null,
+        issueLoadError: null,
+        commandInFlight: false,
+        commandResult: null,
+        events: [],
+        timelineEntries: [],
+        commandCorrelation: null,
+        connectionPhase: "connecting",
+        refreshPhase: "idle",
+        hasSuccessfulRefresh: false,
+        lastRefreshAt: null,
+        showDoneTrackedIssues: false,
+      };
+
+      const elements = {
+        connectionState: document.getElementById("connection-state"),
+        connectionStatePill: document.getElementById("connection-state-pill"),
+        freshnessState: document.getElementById("freshness-state"),
+        refreshState: document.getElementById("refresh-state"),
+        selectedIssueBadge: document.getElementById("selected-issue-badge"),
+        lastRefreshBadge: document.getElementById("last-refresh-badge"),
+        lastRefreshPill: document.getElementById("last-refresh-pill"),
+        loopModeBadge: document.getElementById("loop-mode-badge"),
+        loopStateSummary: document.getElementById("loop-state-summary"),
+        overviewHeadline: document.getElementById("overview-headline"),
+        overviewDetail: document.getElementById("overview-detail"),
+        overviewWarning: document.getElementById("overview-warning"),
+        nextIssueTitle: document.getElementById("next-issue-title"),
+        nextIssueDetail: document.getElementById("next-issue-detail"),
+        primaryActionTitle: document.getElementById("primary-action-title"),
+        primaryActionDetail: document.getElementById("primary-action-detail"),
+        attentionList: document.getElementById("attention-list"),
+        focusBreadcrumb: document.getElementById("focus-breadcrumb"),
+        heroBadgeRow: document.getElementById("hero-badge-row"),
+        heroReason: document.getElementById("hero-reason"),
+        heroPrimaryButton: document.getElementById("hero-primary-button"),
+        heroSecondaryButton: document.getElementById("hero-secondary-button"),
+        heroTertiaryButton: document.getElementById("hero-tertiary-button"),
+        selectedIssueHeading: document.getElementById("selected-issue-heading"),
+        selectedIssueDetail: document.getElementById("selected-issue-detail"),
+        selectedIssueSummaryMetrics: document.getElementById("selected-issue-summary-metrics"),
+        selectedIssueSummaryNotes: document.getElementById("selected-issue-summary-notes"),
+        detailsDisclosure: document.getElementById("details-disclosure"),
+        navOverviewHeading: document.getElementById("nav-overview-heading"),
+        navPanelStatus: document.getElementById("nav-panel-status"),
+        navPanelDoctor: document.getElementById("nav-panel-doctor"),
+        navPanelIssueDetails: document.getElementById("nav-panel-issue-details"),
+        navPanelIssueHistory: document.getElementById("nav-panel-issue-history"),
+        navPanelTrackedHistory: document.getElementById("nav-panel-tracked-history"),
+        navPanelOperatorActions: document.getElementById("nav-panel-operator-actions"),
+        navPanelLiveEvents: document.getElementById("nav-panel-live-events"),
+        navPanelOperatorTimeline: document.getElementById("nav-panel-operator-timeline"),
+        statusReconciliation: document.getElementById("status-reconciliation"),
+        statusLoopChip: document.getElementById("status-loop-chip"),
+        statusMetrics: document.getElementById("status-metrics"),
+        statusWorkflow: document.getElementById("status-workflow"),
+        statusLines: document.getElementById("status-lines"),
+        runtimeRecoverySummary: document.getElementById("runtime-recovery-summary"),
+        runtimeRecoveryLines: document.getElementById("runtime-recovery-lines"),
+        statusPanelWarning: document.getElementById("status-panel-warning"),
+        trackedHistorySummary: document.getElementById("tracked-history-summary"),
+        trackedHistoryLines: document.getElementById("tracked-history-lines"),
+        trackedHistoryToggle: document.getElementById("tracked-history-toggle"),
+        doctorOverall: document.getElementById("doctor-overall"),
+        doctorDecision: document.getElementById("doctor-decision"),
+        doctorTiers: document.getElementById("doctor-tiers"),
+        doctorChecks: document.getElementById("doctor-checks"),
+        issueSummary: document.getElementById("issue-summary"),
+        issueHistorySummary: document.getElementById("issue-history-summary"),
+        issueHistoryLines: document.getElementById("issue-history-lines"),
+        issueMetrics: document.getElementById("issue-metrics"),
+        issueShortcuts: document.getElementById("issue-shortcuts"),
+        issueExplain: document.getElementById("issue-explain"),
+        issueLint: document.getElementById("issue-lint"),
+        issueForm: document.getElementById("issue-form"),
+        issueNumberInput: document.getElementById("issue-number-input"),
+        commandStatus: document.getElementById("command-status"),
+        commandResult: document.getElementById("command-result"),
+        runOnceButton: document.getElementById("run-once-button"),
+        requeueButton: document.getElementById("requeue-button"),
+        pruneWorkspacesButton: document.getElementById("prune-workspaces-button"),
+        resetJsonStateButton: document.getElementById("reset-json-state-button"),
+        operatorTimeline: document.getElementById("operator-timeline"),
+        eventList: document.getElementById("event-list"),
+      };
+
+      const knownEventTypes = [
+        "supervisor.recovery",
+        "supervisor.active_issue.changed",
+        "supervisor.loop.skipped",
+        "supervisor.run_lock.blocked",
+        "supervisor.review_wait.changed",
+      ];
+
+      const COMMAND_CORRELATION_WINDOW_MS = 15000;
+      const mutationAuthStorageKey = ${JSON.stringify(WEBUI_MUTATION_AUTH_STORAGE_KEY)};
+      const mutationAuthHeader = ${JSON.stringify(WEBUI_MUTATION_AUTH_HEADER)};`;
+}
